+++ __tests__/dud.test.js
@@ -1,0 +1,78 @@
+const moment = require('moment');
+const { onMessage } = require('../src/actions');
+const { sendTestMessage } = require('./lib/sendMessageHelper');
+
+const { describeDBSetupTeardown } = require('./lib/dbHelper');
+
+describeDBSetupTeardown();
+moment.locale('ru');
+
+const createMockedSlimbot = (sendMessageFn) => ({ sendMessage: jest.fn(sendMessageFn) });
+
+describe('manual create messages', () => {
+  test('dud', async () => {
+    const slimbot = createMockedSlimbot((chatId, text) => {
+      expect(text).toMatch('*user2* - Дудь (задал 2 вопроса)');
+    });
+
+    await sendTestMessage({
+      userId: 1, firstName: 'user1', date: 1591786800, type: 'text',
+    }, onMessage, slimbot, 'Как дела?');
+
+    await sendTestMessage({
+      userId: 2, firstName: 'user2', date: 1591790400, type: 'text',
+    }, onMessage, slimbot, '?');
+
+    await sendTestMessage({
+      userId: 2, firstName: 'user2', date: 1591790400, type: 'text',
+    }, onMessage, slimbot, 'Норм. А у тебя?');
+
+    const statMessage = {
+      chat: {
+        id: 1,
+      },
+      date: 1591887600,
+      text: '/stats',
+      entities: [{ type: 'bot_command' }],
+    };
+
+    await onMessage(slimbot, statMessage);
+  });
+
+  test('dud five messages', async () => {
+    const slimbot = createMockedSlimbot((chatId, text) => {
+      expect(text).toMatch('*user2* - Дудь (задал 5 вопросов)');
+    });
+
+    await sendTestMessage({
+      userId: 2, firstName: 'user2', date: 1591790400, type: 'text',
+    }, onMessage, slimbot, 'Как дела?');
+
+    await sendTestMessage({
+      userId: 2, firstName: 'user2', date: 1591790400, type: 'text',
+    }, onMessage, slimbot, '?');
+
+    await sendTestMessage({
+      userId: 2, firstName: 'user2', date: 1591790400, type: 'text',
+    }, onMessage, slimbot, 'Как дела? Норм');
+
+    await sendTestMessage({
+      userId: 2, firstName: 'user2', date: 1591790400, type: 'text',
+    }, onMessage, slimbot, '??');
+
+    await sendTestMessage({
+      userId: 2, firstName: 'user2', date: 1591790400, type: 'text',
+    }, onMessage, slimbot, '?');
+
+    const statMessage = {
+      chat: {
+        id: 1,
+      },
+      date: 1591887600,
+      text: '/stats',
+      entities: [{ type: 'bot_command' }],
+    };
+
+    await onMessage(slimbot, statMessage);
+  });
+});
